Use AWS Session Tokens
Back when I wrote this soon-to-be-deprecated FUCKING MONSTROSITY, I was
too lazy to just use STS for authentication.

A few years later, I accidentally committed my admin AK/SK. Oops.

Since then, I made the time to use STS in all of my projects. Turns
out...it's not that hard!